BusinessLogicLayer create user.
Added functionality to BusinessLogicLayer so it can communicate with both the DataAccessLayer and the UsersController.
It uses the UserService to convert models to entities and to validate the entity.
If everything is valid it sends the entity to DataAccessLayer.

Added GetAllAsync to DataAccessLayer to be able to check for users before you save a new one.